fix(ci): line-limit guardrail — split task_tag adoption tests into part_005
The task_tag adoption-replay regression tests pushed
test_pr_monitor_adoption_part_001.py to 1511 lines, over the 1500-line
first-party maintainability limit enforced by
test_first_party_code_files_stay_under_line_limit. Move the two task_tag
policy-conflict tests into a new test_pr_monitor_adoption_part_005.py
(reusing the part_002 scaffolding, mirroring the existing part_004 split),
bringing part_001 back to 1453 lines.
